move helper functions to packages where they used
